Refactoring user admin impl

+++ CommonSQL/src/main/scala/anorm/AnormUtil.scala
@@ -11,20 +11,21 @@ object AnormUtil {
   def isNull(row: Row, columnName: String): Boolean = row.get(columnName).get._1 == null
 
   @deprecated("Create a pull request for anorm instead...")
-  def batchKeys(sql: BatchSql)(implicit conn: Connection): Seq[Long] = {
+  def batchKeys(sql: BatchSql)(implicit conn: Connection): Seq[Int] = {
     
     val stmt = sql.getFilledStatement(conn, true)
     val result = stmt.executeBatch()
+
     val keys = stmt.getGeneratedKeys()
 
     try {
       if (result.exists(_ != 1))
         throw new RuntimeException("Failed batch update")
 
-      val buf = Buffer[Long]()
+      val buf = Buffer[Int]()
 
       while (keys.next()) {
-        buf += keys.getLong("id")
+        buf += keys.getInt("id")
       }
 
       buf
